Complete conceptification of sort.hpp, consolidate concepts into single header
