+++ include/RAJA/pattern/concepts.hpp
@@ -3,7 +3,8 @@
  *
  * \file
  *
- * \brief   Header file defining core C++ concepts for use in development of RAJA
+ * \brief   Header file defining core C++ concepts for use in development of
+ *RAJA
  *
  ******************************************************************************
  */
@@ -64,13 +65,37 @@ template<typename T>
 concept MultiPolicyConcept =
     static_cast<bool>(RAJA::type_traits::is_multi_policy<T>::value);
 
+template<class Function, class Return, class Arg1 = Return, class Arg2 = Arg1>
+concept BinaryFunction = std::is_invocable_r_v<Return, Function&, Arg1, Arg2>;
+
+template<class Function, class Return, class Arg1 = Return>
+concept UnaryFunction = std::is_invocable_r_v<Return, Function&, Arg1>;
+
 }  // namespace concepts
 
 namespace type_traits
 {
 DefineTypeTraitFromConcept(is_execution_policy,
                            RAJA::concepts::ExecutionPolicy);
-}
+
+template<class Function, class Return, class Arg1 = Return, class Arg2 = Arg1>
+struct is_binary_function
+    : std::bool_constant<
+          RAJA::concepts::BinaryFunction<Function, Return, Arg1, Arg2>>
+{};
+template<class Function, class Return, class Arg1 = Return, class Arg2 = Arg1>
+inline constexpr bool is_binary_function_v =
+    is_binary_function<Function, Return, Arg1, Arg2>::value;
+
+template<class Function, class Return, class Arg = Return>
+struct is_unary_function
+    : std::bool_constant<RAJA::concepts::UnaryFunction<Function, Return, Arg>>
+{};
+template<class Function, class Return, class Arg1 = Return>
+inline constexpr bool is_unary_function_v =
+    is_unary_function<Function, Return, Arg1>::value;
+
+}  // namespace type_traits
 
 
 }  // namespace RAJA
