+++ include/RAJA/util/resource.hpp
@@ -23,7 +23,6 @@
 #define RAJA_resource_HPP
 
 #include "camp/resource.hpp"
-#include <type_traits>
 #if defined(RAJA_CUDA_ACTIVE)
 #include "RAJA/policy/cuda/policy.hpp"
 #endif
